Centralize markdown compilation
